refactor: some cleanup application manager
